+++ src/styleUtils.ts
@@ -86,6 +86,10 @@ function mergeMarkdownStyleWithDefault(input: PartialMarkdownStyle | undefined):
   return output;
 }
 
+function parseStyleToNumber(style: string | null): number {
+  return style ? parseInt(style.replace('px', ''), 10) : 0;
+}
+
 export type {PartialMarkdownStyle};
 
-export {mergeMarkdownStyleWithDefault};
+export {mergeMarkdownStyleWithDefault, parseStyleToNumber};
